* Options can now be applied to an arbitrary group of pages. * Support building with MS Visual C++
git-svn-id: https://scantailor.svn.sourceforge.net/svnroot/scantailor/trunk@369 80b70061-164d-0410-b7b7-e274ec932871

+++ filters/select_content/ContentBoxFinder.cpp
@@ -1247,7 +1247,7 @@ ContentBoxFinder::trim(
 		if (num_text_pixels < upper_threshold) {
 			text_influence = min_text_influence +
 					(max_text_influence - min_text_influence)
-					* log(num_text_pixels) / log(upper_threshold);
+					* log((double)num_text_pixels) / log((double)upper_threshold);
 		}
 		//qDebug() << "text_influence = " << text_influence;
 		
@@ -1290,8 +1290,8 @@ ContentBoxFinder::trim(
 	for (int y = removed_area.top(); y <= removed_area.bottom(); ++y) {
 		for (int x = removed_area.left(); x <= removed_area.right(); ++x) {
 			if (cb_line[x >> 5] & (msb >> (x & 31))) {
-				sum_dist_to_garbage += sqrt(dm_garbage_line[x]);
-				sum_dist_to_others += sqrt(dm_others_line[x]);
+				sum_dist_to_garbage += sqrt((double)dm_garbage_line[x]);
+				sum_dist_to_others += sqrt((double)dm_others_line[x]);
 				++count;
 			}
 		}
